+++ app/routes.ts
@@ -3,5 +3,5 @@ import { route, type RouteConfig } from "@react-router/dev/routes";
 export default [
   route("/", "routes/tiku/index.tsx", [route(":textbookId", "routes/tiku/list.tsx"), route(":textbookId/similar", "routes/tiku/similar.tsx")]),
   route("/home", "routes/home/index.tsx"),
-  route("/exam", "routes/exam/index.tsx"),
+  route("/paper", "routes/paper/index.tsx"),
 ] satisfies RouteConfig;
